+++ packages/perseus-core/src/parse-perseus-json/perseus-parsers/interactive-graph-widget.ts
@@ -1,4 +1,3 @@
-import {lockedFigureColorNames} from "../../data-schema";
 import {
     array,
     boolean,
@@ -168,7 +167,6 @@ const parsePerseusGraphTypeVector = object({
     coords: optional(nullable(pair(pairOfNumbers, pairOfNumbers))),
     startCoords: optional(pair(pairOfNumbers, pairOfNumbers)),
     match: optional(enumeration("exact", "congruent")),
-    pointLabels: optional(pair(string, string)),
 });
 
 export const parsePerseusGraphType = discriminatedUnionOn("type")
@@ -189,7 +187,20 @@ export const parsePerseusGraphType = discriminatedUnionOn("type")
     .withBranch("logarithm", parsePerseusGraphTypeLogarithm)
     .withBranch("vector", parsePerseusGraphTypeVector).parser;
 
-const parseLockedFigureColor = enumeration(...lockedFigureColorNames);
+const parseLockedFigureColor = pipeParsers(
+    enumeration(
+        // Same as lockedFigureColorNames in data-schema.ts
+        "blue",
+        "gold",
+        "green",
+        "grayH",
+        "purple",
+        "pink",
+        "red",
+        // deprecated name - "orange" is now "gold"
+        "orange",
+    ),
+).then(convert((color) => (color === "orange" ? "gold" : color))).parser;
 
 const parseLockedFigureFillType = enumeration(
     "none",