+++ packages/perseus/src/widgets/interactive-graphs/graphs/point.tsx
@@ -1,12 +1,14 @@
 import {useTimeout} from "@khanacademy/wonder-blocks-timing";
 import * as React from "react";
 
-import {usePerseusI18n} from "../../../components/i18n-context";
 import {actions} from "../reducer/interactive-graph-action";
 import useGraphConfig from "../reducer/use-graph-config";
 import {getCSSZoomFactor} from "../utils";
 
-import {buildPointAriaLabel} from "./components/build-point-aria-label";
+import {
+    buildPointAriaLabel,
+    usePointAriaLabel,
+} from "./components/build-point-aria-label";
 import {MovablePoint} from "./components/movable-point";
 import {srFormatNumber} from "./screenreader-text";
 import {useTransformVectorsToPixels, pixelsToVectors} from "./use-transform";
@@ -82,7 +84,7 @@ function PointGraph(props: Props) {
 function LimitedPointGraph(statefulProps: StatefulProps) {
     const {dispatch} = statefulProps;
     const {pointLabels} = statefulProps.graphState;
-    const {strings, locale} = usePerseusI18n();
+    const buildLabel = usePointAriaLabel(pointLabels);
 
     return (
         <>
@@ -97,13 +99,7 @@ function LimitedPointGraph(statefulProps: StatefulProps) {
                     // TODO(LEMS-4189): Remove ariaLive once aria-live is
                     // dropped from useControlPoint.
                     ariaLive="off"
-                    ariaLabel={buildPointAriaLabel(
-                        pointLabels,
-                        i,
-                        point,
-                        strings,
-                        locale,
-                    )}
+                    ariaLabel={buildLabel(i, point)}
                     onMove={(destination) =>
                         dispatch(actions.pointGraph.movePoint(i, destination))
                     }
@@ -116,7 +112,7 @@ function LimitedPointGraph(statefulProps: StatefulProps) {
 function UnlimitedPointGraph(statefulProps: StatefulProps) {
     const {dispatch, graphConfig, pointsRef, top, left} = statefulProps;
     const {coords, pointLabels} = statefulProps.graphState;
-    const {strings, locale} = usePerseusI18n();
+    const buildLabel = usePointAriaLabel(pointLabels);
 
     // When users drag a point on iOS Safari, the browser fires a click event after the mouseup
     // at the original click location, which would add an unwanted new point. We track drag
@@ -178,13 +174,7 @@ function UnlimitedPointGraph(statefulProps: StatefulProps) {
                     // TODO(LEMS-4189): Remove ariaLive once aria-live is
                     // dropped from useControlPoint.
                     ariaLive="off"
-                    ariaLabel={buildPointAriaLabel(
-                        pointLabels,
-                        i,
-                        point,
-                        strings,
-                        locale,
-                    )}
+                    ariaLabel={buildLabel(i, point)}
                     onDragStart={() => {
                         dragEndCallbackTimer.clear();
                         setIsCurrentlyDragging(true);